Toggle RGB with caps lock
Keychron Q6 has a capslock key but it's not very visible if the keycap isn't backlit.
Get around this by toggling the RGB lights for the whole keyboard for increased visibility.

If the RGB lights are enabled when the keyboard loses power, they'll stay enabled when power is returned but the capslock will be disabled.
This is prevented by using matrix_scan_user.

In the case that this state is ever encountered (rgb lights on, caps lock off) then an additional bool is used for the rgb lighting to enable/disable appropriately